+++ src/components/dashboard/item/item_page.tsx
@@ -110,21 +110,33 @@ export function ItemPage(): React.JSX.Element {
   }
 
   const handleDeleteitem = (id: number) => {
-    fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/item/${id}`, {
-      method: 'DELETE',
-      headers: {
-        'Authorization': `Bearer ${authClient.getToken()}`
+    Swal.fire({
+      title: 'please confirm',
+      text: 'Are you sure you want to delete the item?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/item/${id}`, {
+          method: 'DELETE',
+          headers: {
+            'Authorization': `Bearer ${authClient.getToken()}`
+          }
+        }).then((res) => {
+          if (res.ok) {
+            fetchitemData();
+            toast.success("item deleted successfully");
+          } else {
+            throw new Error("Failed to delete item");
+          }
+        }).catch((err) => {
+          toast.error("Failed to delete item");
+        });
       }
-    }).then((res) => {
-      if (res.ok) {
-        fetchitemData();
-        toast.success("item deleted successfully");
-      } else {
-        throw new Error("Failed to delete item");
-      }
-    }).catch((err) => {
-      toast.error("Failed to delete item");
-    });
+    })
   }
 
   const handleRemoveEngineer = (id: number) => {
@@ -176,7 +188,7 @@ export function ItemPage(): React.JSX.Element {
 
   return (
     <>
-      <EngineerSelectionModal open={eopen} handleClose={handleEClose} itemID={itemID} fetchitemData={fetchitemData}/>
+      <EngineerSelectionModal open={eopen} handleClose={handleEClose} itemID={itemID} fetchitemData={fetchitemData} />
       <ItemModalForm open={open} handleClose={handleModalClose} itemID={itemID} fetchitemData={fetchitemData} />
       <Stack direction="row" spacing={3}>
         <Stack spacing={1} sx={{ flex: '1 1 auto' }}>
@@ -229,7 +241,7 @@ export function ItemPage(): React.JSX.Element {
                     <TableCell>{row.category.name}</TableCell>
                     <TableCell>{row.brand.name}</TableCell>
                     <TableCell>{row.model.name}</TableCell>
-                    <TableCell sx={{textAlign: 'center'}}><Badge badgeContent={row.engineer?.name ? row.engineer.name : 'WareHouse'} color={row.engineer?.name ? 'primary' : 'warning'} /></TableCell>
+                    <TableCell sx={{ textAlign: 'center' }}><Badge badgeContent={row.engineer?.name ? row.engineer.name : 'WareHouse'} color={row.engineer?.name ? 'primary' : 'warning'} /></TableCell>
                     <TableCell>{dayjs(row.waranty_expiry_date).format('MMM D, YYYY')}</TableCell>
                     <TableCell><Badge badgeContent={row.status} color={row.status === 'in_stock' ? 'success' : 'warning'} /></TableCell>
                     {/* <TableCell>{dayjs(row.created_at).format('MMM D, YYYY')}</TableCell> */}
